refactor: 보너스 번호 클래스화
